% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scSVA.R
\name{scSVA}
\alias{scSVA}
\title{Run scSVA}
\usage{
scSVA(background = FALSE)
}
\arguments{
\item{background}{Whether to run Shiny App in background (access to the console while running scSVA), default=FALSE}
}
\value{
Open Shiny App
}
\description{
Run scSVA
}
\examples{
scSVA()
}
